fix: Add button and maximize button fixes
