+++ app/client/src/pages/Editor/IDE/EditorPane/UI/Add.tsx
@@ -1,13 +1,18 @@
 import React, { useCallback } from "react";
-import { Flex } from "design-system";
+import { Button, Flex, Text } from "design-system";
 import { useSelector } from "react-redux";
+import styled from "styled-components";
 
 import history from "utils/history";
 import { getCurrentPageId } from "@appsmith/selectors/entitiesSelector";
 import UIEntitySidebar from "pages/Editor/widgetSidebar/UIEntitySidebar";
 import { widgetListURL } from "@appsmith/RouteBuilder";
-import { EDITOR_PANE_TEXTS } from "@appsmith/constants/messages";
-import SegmentAddHeader from "../components/SegmentAddHeader";
+import { EDITOR_PANE_TEXTS, createMessage } from "@appsmith/constants/messages";
+
+const Container = styled(Flex)`
+  padding-right: var(--ads-v2-spaces-2);
+  background-color: var(--ads-v2-color-gray-50);
+`;
 
 const AddWidgets = (props: { focusSearchInput?: boolean }) => {
   const pageId = useSelector(getCurrentPageId) as string;
@@ -18,10 +23,26 @@ const AddWidgets = (props: { focusSearchInput?: boolean }) => {
 
   return (
     <>
-      <SegmentAddHeader
-        onCloseClick={closeButtonClickHandler}
-        titleMessage={EDITOR_PANE_TEXTS.widgets_create_tab_title}
-      />
+      <Container
+        alignItems="center"
+        borderBottom={"1px solid var(--ads-v2-color-border)"}
+        justifyContent="space-between"
+        px="spaces-4"
+        py="spaces-2"
+      >
+        <Text color="var(--ads-v2-color-fg)" kind="heading-xs">
+          {createMessage(EDITOR_PANE_TEXTS.widgets_create_tab_title)}
+        </Text>
+        <Button
+          aria-label="Close pane"
+          data-testid="t--add-pane-close-icon"
+          isIconButton
+          kind={"tertiary"}
+          onClick={closeButtonClickHandler}
+          size={"sm"}
+          startIcon={"close-line"}
+        />
+      </Container>
       <Flex flexDirection="column" gap="spaces-3" overflowX="scroll">
         <UIEntitySidebar focusSearchInput={props.focusSearchInput} isActive />
       </Flex>
