+++ app/client/src/pages/Editor/IDE/EditorTabs/SplitScreenTabs.tsx
@@ -1,7 +1,7 @@
-import React from "react";
+import React, { useCallback } from "react";
 
 import FileTabs from "./FileTabs";
-import { useSelector } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import { getIDEViewMode, getIsSideBySideEnabled } from "selectors/ideSelectors";
 import Container from "./Container";
 import { useCurrentEditorState, useIDETabClickHandlers } from "../hooks";
@@ -13,8 +13,16 @@ import { TabSelectors } from "./constants";
 import { Announcement } from "../EditorPane/components/Announcement";
 import { SearchableFilesList } from "./SearchableFilesList";
 import { AddButton } from "./AddButton";
+import { Button, Tooltip } from "design-system";
+import {
+  MAXIMIZE_BUTTON_TOOLTIP,
+  createMessage,
+} from "@appsmith/constants/messages";
+import AnalyticsUtil from "@appsmith/utils/AnalyticsUtil";
+import { setIdeEditorViewMode } from "actions/ideActions";
 
 const SplitScreenTabs = () => {
+  const dispatch = useDispatch();
   const isSideBySideEnabled = useSelector(getIsSideBySideEnabled);
   const ideViewMode = useSelector(getIDEViewMode);
   const { segment } = useCurrentEditorState();
@@ -25,26 +33,44 @@ const SplitScreenTabs = () => {
   const files = useSelector(tabsConfig.tabsSelector);
   const allFilesList = useSelector(tabsConfig.listSelector);
 
+  const handleMaximizeButtonClick = useCallback(() => {
+    AnalyticsUtil.logEvent("EDITOR_MODE_CHANGE", {
+      to: EditorViewMode.FullScreen,
+    });
+    dispatch(setIdeEditorViewMode(EditorViewMode.FullScreen));
+  }, []);
+
   if (!isSideBySideEnabled) return null;
   if (ideViewMode === EditorViewMode.FullScreen) return null;
   if (segment === EditorEntityTab.UI) return null;
 
   return (
     <>
-      {files.length > 0 ? (
-        <Container>
-          <SearchableFilesList
-            allItems={allFilesList}
-            navigateToTab={tabClickHandler}
+      {/* {files.length > 0 ? ( */}
+      <Container>
+        <SearchableFilesList
+          allItems={allFilesList}
+          navigateToTab={tabClickHandler}
+        />
+        <FileTabs
+          navigateToTab={tabClickHandler}
+          onClose={closeClickHandler}
+          tabs={files}
+        />
+        {files.length > 0 ? <AddButton /> : null}
+
+        <Tooltip content={createMessage(MAXIMIZE_BUTTON_TOOLTIP)}>
+          <Button
+            className="ml-auto min-w-[24px]"
+            data-testid="t--ide-maximize"
+            id="editor-mode-maximize"
+            isIconButton
+            kind="tertiary"
+            onClick={handleMaximizeButtonClick}
+            startIcon="maximize-v3"
           />
-          <FileTabs
-            navigateToTab={tabClickHandler}
-            onClose={closeClickHandler}
-            tabs={files}
-          />
-          <AddButton />
-        </Container>
-      ) : null}
+        </Tooltip>
+      </Container>
       <Announcement />
     </>
   );
